+++ src/lib.rs
@@ -61,7 +61,7 @@ pub use version::{API_SPEC_VERSION, CLIENT_VERSION};
 pub use clients::actor::{ActorBuildOptions, ActorStartOptions};
 pub use clients::actor_collection::ActorListOptions;
 pub use clients::dataset::{DatasetDownloadOptions, DatasetListItemsOptions, DownloadItemsFormat};
-pub use clients::key_value_store::{GetRecordOptions, ListKeysOptions};
+pub use clients::key_value_store::{GetRecordOptions, GetRecordsOptions, ListKeysOptions};
 pub use clients::request_queue::ListRequestsOptions;
 pub use clients::run::RunResurrectOptions;
 pub use clients::run_collection::RunListOptions;